+++ src/21_scoringInterface.js
@@ -16,8 +16,8 @@ function addSetOfScoringSteps(SS, sheetModeID, Config, IndicatorsObj, ResearchSt
 
     // var to estimate max sheet width in terms of columns based on whether G has subcomponents. This is needed for formatting the whole sheet at end of script. More performant than using getLastCol() esp. when executed per Sheet (think 45 indicators)
     var globalNrOfComponents = 1
-    if (IndicatorsObj.indicatorClasses[0].components) {
-        globalNrOfComponents = IndicatorsObj.indicatorClasses[0].components.length
+    if (IndicatorsObj.indicatorCategories[0].components) {
+        globalNrOfComponents = IndicatorsObj.indicatorCategories[0].components.length
     }
 
     var numberOfColumns = (CompanyObj.numberOfServices + 2) * globalNrOfComponents + 1