+++ src/40_feedbackMain.js
@@ -24,7 +24,7 @@ function createFeedbackForms(useIndicatorSubset, thisCompany, filenamePrefix, fi
     var spreadsheetName = spreadSheetFileName(filenamePrefix, mainSheetMode, companyFilename, filenameSuffix)
 
     // connect to Spreadsheet if it already exists (Danger!), otherwise create and return new file
-    var SS = connectToSpreadsheetByName(spreadsheetName, true)
+    var SS = createSpreadsheet(spreadsheetName, true)
     var fileID = SS.getId()
 
     // --- // Feedback Parameters // --- //
@@ -41,8 +41,8 @@ function createFeedbackForms(useIndicatorSubset, thisCompany, filenamePrefix, fi
 
     // var to estimate max sheet width in terms of columns based on whether G has subcomponents. This is needed for formatting the whole sheet at end of script. More performant than using getLastCol() esp. when executed per Sheet (think 45 indicators)
     var globalNrOfComponents = 1
-    if (IndicatorsObj.indicatorClasses[0].components) {
-        globalNrOfComponents = IndicatorsObj.indicatorClasses[0].components.length
+    if (IndicatorsObj.indicatorCategories[0].components) {
+        globalNrOfComponents = IndicatorsObj.indicatorCategories[0].components.length
     }
 
     var companyCols = 1
@@ -71,11 +71,11 @@ function createFeedbackForms(useIndicatorSubset, thisCompany, filenamePrefix, fi
 
     var sourcesSheet = importSourcesSheet(SS, sheetName, CompanyObj, doOverwrite)
 
-    for (var c = 0; c < IndicatorsObj.indicatorClasses.length; c++) {
+    for (var c = 0; c < IndicatorsObj.indicatorCategories.length; c++) {
 
         lastCol = 1
 
-        thisIndClass = IndicatorsObj.indicatorClasses[c]
+        thisIndClass = IndicatorsObj.indicatorCategories[c]
 
         sheetName = thisIndClass.labelLong