+++ src/99_SpreadsheetOperations.js
@@ -1,4 +1,15 @@
-// --- connection helper function  --- //
+// --- File-level connection helper function  --- //
+
+/* global
+    rootFolderID,
+    outputFolderName,
+    assignFolderEditors,
+    assignFolderOwner,
+    assignFileEditors,
+    assignFileOwner,
+    createNewFolder,
+    printParentFolders
+*/
 
 // Connect by Spreadsheet name //
 // Universal Method: Connect to Spreadsheet File and return Obj as open Spreadsheet
@@ -7,11 +18,11 @@
 
 function mainTestConnectionByName() {
     var spreadsheetName = "verizon test"
-    connectToSpreadsheetByName(spreadsheetName, false)
+    createSpreadsheet(spreadsheetName, false)
 }
 
 function mainTestConnectionByID() {
-    connectToSpreadsheetByID(spreadsheetID)
+    openSpreadsheetByID(spreadsheetID)
 }
 
 // TODO: refactor everywhere into two seperate functions:
@@ -19,11 +30,11 @@ function mainTestConnectionByID() {
 // one to create a new file
 // --> separation of concerns AND explicit action where needed
 
-function connectToSpreadsheetByName(spreadsheetName, createNewFile) {
+function createSpreadsheet(spreadsheetName, createNewFile) {
 
-    var Spreadsheets = DriveApp.getFilesByName(spreadsheetName)
+    let Spreadsheets = DriveApp.getFilesByName(spreadsheetName)
 
-    var Spreadsheet = null
+    let SS = null
 
     // if SS doesn't exist
     // --- and createNewFile === true
@@ -37,11 +48,12 @@ function connectToSpreadsheetByName(spreadsheetName, createNewFile) {
         Logger.log(spreadsheetName + " does not exist!")
 
         if (createNewFile) {
-            Logger.log("--- --- START: creating " + spreadsheetName)
+            Logger.log("--- --- START: creating " + spreadsheetName + " in " + rootFolderID + "/" + outputFolderName)
 
-            var folderID = createFolderIfNotExist(rootFolderID, outputFolderName)
+            // TODO: add tests whether script (user?) can access folders
+            let folderID = createNewFolder(rootFolderID, outputFolderName)
 
-            var resource = {
+            let resource = {
                 title: spreadsheetName,
                 mimeType: MimeType.GOOGLE_SHEETS,
                 parents: [{
@@ -50,51 +62,56 @@ function connectToSpreadsheetByName(spreadsheetName, createNewFile) {
             }
 
             // Logger.log(resource.parents.id)
-            var fileJson = Drive.Files.insert(resource)
+            let fileJson = Drive.Files.insert(resource)
 
-            var fileId = fileJson.id
+            let fileId = fileJson.id
+
+            let File = DriveApp.getFileById(fileId)
+            File.setOwner("data@rankingdigitalrights.org")
+            let path = printParentFolders(File)
             Logger.log("new Speadsheet fileID: " + fileId)
-            Spreadsheet = connectToSpreadsheetByID(fileId)
+            Logger.log("File Path: " + path)
+            SS = openSpreadsheetByID(fileId)
 
         } else {
-            Spreadsheet = null
+            SS = null
+            Logger.log(spreadsheetName + " does not exist and NOT creating a new file")
         }
 
     } else {
 
-        // while (Spreadsheet.hasNext()) { }
-        // Nope. Only do for first Spreadsheet element
-        var thisSpreadsheet = Spreadsheets.next()
-        Logger.log("File " + thisSpreadsheet.getName() + " exists")
-        Logger.log("locally connected to: " + thisSpreadsheet.getName())
+        // Only do for first Spreadsheet element
+        SS = Spreadsheets.next()
+        Logger.log("File " + SS.getName() + " exists")
+        Logger.log("locally connected to: " + SS.getName())
 
-        Spreadsheet = SpreadsheetApp.open(thisSpreadsheet)
+        SS = SpreadsheetApp.open(SS)
     }
     // returns SS or null
-    return Spreadsheet
+    return SS
 }
 
 // connect by Spreadsheet ID //
 // more accurate then by name //
 
-function connectToSpreadsheetByID(ID) {
+function openSpreadsheetByID(ID) {
 
-    var thisSpreadsheet = SpreadsheetApp.openById(ID)
-    Logger.log("locally connected to: " + thisSpreadsheet.getName())
-    return thisSpreadsheet
+    let SS = SpreadsheetApp.openById(ID)
+    Logger.log("locally connected to: " + SS.getName())
+    return SS
 
 }
 
 
-// Help Function to overwrite Sheet in Spreadsheet if it is already existing
+// Helper Function to overwrite Sheet in Spreadsheet if it is already existing
 
-function insertSheetIfNotExist(Spreadsheet, SheetName, updateSheet) {
-    var Sheet
-    if (!Spreadsheet.getSheetByName(SheetName)) {
-        Sheet = Spreadsheet.insertSheet(SheetName)
+function insertSheetIfNotExist(SS, SheetName, updateSheet) {
+    let Sheet
+    if (!SS.getSheetByName(SheetName)) {
+        Sheet = SS.insertSheet(SheetName)
     } else {
         if (updateSheet) {
-            Sheet = Spreadsheet.getSheetByName(SheetName)
+            Sheet = SS.getSheetByName(SheetName)
         } else {
             Sheet = null
             Logger.log("WARN: " + "Sheet for " + SheetName + " already exists ")
@@ -103,50 +120,37 @@ function insertSheetIfNotExist(Spreadsheet, SheetName, updateSheet) {
     return Sheet
 }
 
-function moveHideSheetifExists(Spreadsheet, Sheet, posInt) {
+function moveHideSheetifExists(SS, Sheet, posInt) {
 
     if (!posInt) {
         posInt = 1
     }
 
     if (Sheet !== null) {
-        moveSheetToPos(Spreadsheet, Sheet, posInt)
+        moveSheetToPos(SS, Sheet, posInt)
         Sheet.hideSheet()
     }
 }
 
-function moveSheetifExists(Spreadsheet, Sheet, posInt) {
+function moveSheetifExists(SS, Sheet, posInt) {
 
     if (!posInt) {
         posInt = 1
     }
 
     if (Sheet !== null) {
-        moveSheetToPos(Spreadsheet, Sheet, posInt)
+        moveSheetToPos(SS, Sheet, posInt)
     }
 }
 
-
-
-function moveSheetToPos(Spreadsheet, Sheet, posInt) {
-    Spreadsheet.setActiveSheet(Sheet)
-    Spreadsheet.moveActiveSheet(posInt)
-}
-
-function addFileIDtoControl(mode, companyShortName, fileID, controlSpreadsheetID) {
-
-    var spreadsheet = connectToSpreadsheetByID(controlSpreadsheetID)
-    var sheet = insertSheetIfNotExist(spreadsheet, mode, true)
-    var formula = "=HYPERLINK(CONCAT(\"https://docs.google.com/spreadsheets/d/\",INDIRECT(ADDRESS(ROW(),COLUMN()-1))),INDIRECT(ADDRESS(ROW(),COLUMN()-2)))"
-    sheet.appendRow([mode, companyShortName, fileID, formula])
-    Logger.log("created" + fileID + "; added to Control")
-
+function moveSheetToPos(SS, Sheet, posInt) {
+    SS.setActiveSheet(Sheet)
+    SS.moveActiveSheet(posInt)
 }
 
 function importRangeFormula(url, range, integrateOutputs) {
 
-    var formula
-
+    let formula
     if (integrateOutputs) {
         formula = "=" + range
     } else {
@@ -156,19 +160,20 @@ function importRangeFormula(url, range, integrateOutputs) {
     return formula
 }
 
-function getSheetByName(Spreadsheet, Sheetname) {
-    var Sheet
-    if (!Spreadsheet.getSheetByName(Sheetname)) {
+// Sheets have unique Names, so no iteration
+function getSheetByName(SS, Sheetname) {
+    let Sheet
+    if (!SS.getSheetByName(Sheetname)) {
         Sheet = null
         Logger.log("Sheet " + Sheetname + " not found.")
     } else {
-        Sheet = Spreadsheet.getSheetByName(Sheetname)
+        Sheet = SS.getSheetByName(Sheetname)
     }
     return Sheet
 }
 
 function removeEmptySheet(SS) {
-    var emptySheet = SS.getSheetByName("Sheet1")
+    let emptySheet = SS.getSheetByName("Sheet1")
 
     if (emptySheet) {
         SS.deleteSheet(emptySheet)
@@ -176,9 +181,24 @@ function removeEmptySheet(SS) {
 }
 
 function resizeSheet(Sheet, newRows) {
-    var oldRows = Sheet.getMaxRows()
-    var rowDiff = newRows - oldRows
+    let oldRows = Sheet.getMaxRows()
+    let rowDiff = newRows - oldRows
     if (oldRows < newRows) {
         Sheet.insertRows(1, rowDiff)
     }
+}
+
+function isValueInColumn(SS, columnNr, mode, value) {
+    let Range, Sheet, lastRow
+    let isInColumn = false
+    Sheet = SS.getSheetByName(mode)
+    lastRow = Sheet.getLastRow()
+    Logger.log("lastRow: " + lastRow)
+    if (lastRow >= 1) {
+        Range = Sheet.getRange(1, columnNr, lastRow)
+        isInColumn = Range.getValues()
+            .flat(2) // unnest 2D array to flat 1D array
+            .includes(value)
+    }
+    return isInColumn
 }