+++ toys/Test.js
@@ -5,6 +5,6 @@
 function testSomething() {
 
     var testEffect = centralConfig.indexPrefix
-    //  var testEffect = connectToSpreadsheetByName("2019 Pilot - Amazon DC Testing", false)
+    //  var testEffect = createSpreadsheet("2019 Pilot - Amazon DC Testing", false)
     Logger.log(testEffect)
 }